Ajustando path para concern
