+++ app/controllers/api/v1/stores_controller.rb
@@ -1,3 +1,3 @@
 class Api::V1::StoresController < ApplicationController
-  include Api::V1::Concerns::StoreActions
+  include Api::V1::StoreActions
 end
